install-dependencies checks if this is an IS

+++ install-dependencies.sh
@@ -4,6 +4,14 @@ set -o errexit
 
 [ ! -e README.md ] && echo Run this from the top-level directory && exit 1
 
+. .scripts/get-config.sh
+
+
+if [[ ! "$AMWA_ID" =~ "IS-" ]]; then
+    echo Nothing to install
+    exit 0
+fi
+
 rm -rf raml2html-nmos-theme
 git clone https://${GITHUB_TOKEN:+${GITHUB_TOKEN}@}github.com/AMWA-TV/raml2html-nmos-theme
 cp .scripts/package.json .
